tests/server-test: fix the tests on usr-merged systems
The testsuite is failing on Ubuntu builders as they operate with /bin a
symlink to /usr/bin. As a result, depending on how you resolve it, `ls`
can either be `/bin/ls` or `/usr/bin/ls`. Since Debian also seems to
transition to such a setup, it might be wise to simply relax the tests.